requirements.txt
# onnx
pytest
onnx
onnxruntime-gpu
supervision
transformers # from tokenizers import Tokenizer

# flask
Flask
Flask-SQLAlchemy
pymysql
Flask-Redis
nacos-sdk-python
PyYAML
snowflake-id  # 雪花算法生成 ID
Flask-APScheduler # 定时任务

# multi-thread
gevent
gunicorn

# json
pydantic